Introduce dynamic logger creation inside subclasses
After the change in the previous commit, the effect of having dynamic logger
creation is not too obvious on the graph. Hence, I have pushed the boundary
and introduce the dynamic logger creation inside the subclass under ImageProcessor.
The changes will force a new logger and a S3Client created for every image
transformation operation.